remove training files, fold other functions into update_ad_types and new class AdCreativeBodyProcessor
