+++ ad_type_classification/update_ad_types.py
@@ -4,49 +4,76 @@ from collections import defaultdict
 import json
 import logging
 import os.path
+import re
 import sys
+from urllib import parse
 
-import pandas
 import joblib
+import pandas
 
 import config_utils
 import db_functions
-from ad_type_classification.helper_fns import (get_normalized_creative_url,
-                                               get_canonical_url_to_ad_type_table)
-from ad_type_classification.text_process_fns import process_creative_body
+from ad_type_classification import ad_creative_body_processor
 
 DATA_DIR = os.path.join('ad_type_classification', 'data')
+URL_RE = re.compile(
+        r'http[s]?://(?:[a-zA-Z]|[0-9]|[$\-_@.&+]|[!*\(\),]|(?:%[0-9a-fA-F][0-9a-fA-F]))+')
 
-def main(config_file_path):
-    """Perform ad classification.
+
+def _get_canonical_url(url):
+    """Parse string for URL, and strip www. if present.
 
     Args:
-        config_file_path: str path to config file.
+        url: str of URL to parse.
+    Returns:
+        str of URL (with www. stripped if present) of parsed URL. empty string if parsing failed.
     """
-    config = config_utils.get_config(config_file_path)
-    with config_utils.get_database_connection_from_config(config) as database_connection:
+    try:
+        parsed_url = parse.urlparse(str(url))
+        domain = str(parsed_url.netloc).lower()
+        path = str(parsed_url.path).lower()
+        if domain.startswith('www.'):
+            domain = domain[4:]
+        return domain + path
+    except ValueError as err:
+        logging.info('%s error while processing url: %s', err, url)
+        return ''
 
-        classifier = joblib.load(os.path.join(DATA_DIR, 'ad_type_classifier.pk1'))
-        label_encoder = joblib.load(os.path.join(DATA_DIR, 'ad_type_label_encoder.pk1'))
-        data_prep_pipeline = joblib.load(os.path.join(DATA_DIR, 'data_prep_pipeline.pk1'))
-        db_interface = db_functions.DBInterface(database_connection)
 
-        ad_type_to_archive_ids = defaultdict(list)
-        canonical_url_to_ad_type_table = get_canonical_url_to_ad_type_table(os.path.join(
-            DATA_DIR, 'ad_url_to_type.csv'))
-        ad_iterator = classify_ad_iterator(
-            db_interface.all_ads_with_nonempty_link_caption_or_body(),
-            canonical_url_to_ad_type_table, classifier,
-            label_encoder, data_prep_pipeline)
-        for classified_ad in ad_iterator:
-            ad_type_to_archive_ids[classified_ad['ad_type']].append(classified_ad['archive_id'])
-        type_sums = ['%s: %d' % (k, len(v)) for k, v in ad_type_to_archive_ids.items()]
-        logging.info('Type sums:\n%s', '\n'.join(type_sums))
-        with open('ad_to_type_mappings.json', 'w') as ad_mapping_file:
-            json.dump(ad_type_to_archive_ids, ad_mapping_file)
+def get_normalized_creative_url(row):
+    """Get URL from ad creative data. Either link caption or parsed from creative body.
 
-        db_interface.update_ad_types(archive_id_and_ad_type_iterator(ad_type_to_archive_ids))
-        database_connection.commit()
+    Args:
+        row: dict with keys 'ad_creative_link_caption' and 'ad_creative_body'.
+    Returns:
+        string: domain (ie secure.actblue.com) of link in ad creative.
+    """
+    url = ''
+    if row.get('ad_creative_link_caption'):
+        url = row['ad_creative_link_caption']
+    else:
+        match = re.search(URL_RE, row['ad_creative_body'])
+        if match:
+            url = _get_canonical_url(match.group(0))
+    return url
+
+
+def get_canonical_url_to_ad_type_table(canonical_url_to_type_csv_path):
+    """Returns dict of canonical normalized_url -> ad_type.
+
+    Args:
+        str path to CSV of ad url and ad type data.
+    Returns:
+        dict normalized_url -> ad_type.
+    """
+    canonical_url_to_type_df = pandas.read_csv(canonical_url_to_type_csv_path)
+    canonical_url_to_type_df['ad_type'] = canonical_url_to_type_df['ad_type'].astype('category')
+    df_as_dicts = canonical_url_to_type_df.to_dict(orient='records')
+    canonical_url_to_ad_type_table = {}
+    for row in df_as_dicts:
+        canonical_url_to_ad_type_table[row['normalized_url']] = row['ad_type']
+    canonical_url_to_ad_type_table[''] = 'INFORM'
+    return canonical_url_to_ad_type_table
 
 
 def classify_ad_iterator(ads, canonical_url_to_ad_type_table, classifier, label_encoder,
@@ -62,6 +89,7 @@ def classify_ad_iterator(ads, canonical_url_to_ad_type_table, classifier, label_
     Yields:
         dict of classified add with keys 'archive_id' and 'ad_type'.
     """
+    processor = ad_creative_body_processor.AdCreativeBodyProcessor()
     to_classify = []
     num_rows_processed = 0
     for result in ads:
@@ -79,7 +107,7 @@ def classify_ad_iterator(ads, canonical_url_to_ad_type_table, classifier, label_
                          len(to_classify), num_rows_processed)
             classification_df = pandas.DataFrame(to_classify)
             classification_df['processed_body'] = classification_df['ad_creative_body'].apply(
-                process_creative_body)
+                processor.process_creative_body)
             prediction_data = data_prep_pipeline.transform(
                 classification_df['processed_body'].astype('str'))
             classification_df['ad_type'] = classifier.predict(prediction_data)
@@ -102,6 +130,39 @@ def archive_id_and_ad_type_iterator(ad_type_to_archive_ids):
     for ad_type in ad_type_to_archive_ids:
         for archive_id in ad_type_to_archive_ids[ad_type]:
             yield (archive_id, ad_type)
+
+
+def main(config_file_path):
+    """Perform ad classification.
+
+    Args:
+        config_file_path: str path to config file.
+    """
+    config = config_utils.get_config(config_file_path)
+    classifier = joblib.load(os.path.join(DATA_DIR, 'ad_type_classifier.pk1'))
+    label_encoder = joblib.load(os.path.join(DATA_DIR, 'ad_type_label_encoder.pk1'))
+    data_prep_pipeline = joblib.load(os.path.join(DATA_DIR, 'data_prep_pipeline.pk1'))
+
+    with config_utils.get_database_connection_from_config(config) as database_connection:
+
+        db_interface = db_functions.DBInterface(database_connection)
+
+        ad_type_to_archive_ids = defaultdict(list)
+        canonical_url_to_ad_type_table = get_canonical_url_to_ad_type_table(os.path.join(
+            DATA_DIR, 'ad_url_to_type.csv'))
+        ad_iterator = classify_ad_iterator(
+            db_interface.all_ads_with_nonempty_link_caption_or_body(),
+            canonical_url_to_ad_type_table, classifier,
+            label_encoder, data_prep_pipeline)
+        for classified_ad in ad_iterator:
+            ad_type_to_archive_ids[classified_ad['ad_type']].append(classified_ad['archive_id'])
+        type_sums = ['%s: %d' % (k, len(v)) for k, v in ad_type_to_archive_ids.items()]
+        logging.info('Type sums:\n%s', '\n'.join(type_sums))
+        with open('ad_to_type_mappings.json', 'w') as ad_mapping_file:
+            json.dump(ad_type_to_archive_ids, ad_mapping_file)
+
+        db_interface.update_ad_types(archive_id_and_ad_type_iterator(ad_type_to_archive_ids))
+        database_connection.commit()
 
 
 if __name__ == '__main__':
